+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 # Version 16
 
 ### [Current changes]
+- **[Add]** New "Rust" based performance action.
 - **[Add]** Added - Text to Speech action
 - **[Add]** Added - S3 Actions
 - **[Fix]** Fix chrome. firefox & edge driver issues on selenium 4.10 (executable_path error).